Move all mathematical functions to separate subpackage
