+++ docs/source/conf.py
@@ -389,4 +389,8 @@ exclude_patterns.append('_autoapi_templates/python/module.rst')
 # Ignore sphinx-autoapi warnings on multiple target description
 suppress_warnings.append('ref.python')
 
+# Ignore sphinx-autoapi warnings on reimported objects
+# See https://github.com/readthedocs/sphinx-autoapi/issues/285
+suppress_warnings.append('autoapi.python_import_resolution')
+
 myst_update_mathjax = False
