+++ src/poliastro/_math/integrate.py
@@ -1,0 +1,3 @@
+from scipy.integrate import quad
+
+__all__ = ["quad"]
