+++ src/poliastro/_math/ivp.py
@@ -1,0 +1,3 @@
+from scipy.integrate import DOP853, solve_ivp
+
+__all__ = ["DOP853", "solve_ivp"]
