+++ src/poliastro/_math/optimize.py
@@ -1,0 +1,3 @@
+from scipy.optimize import brentq
+
+__all__ = ["brentq"]
